feat(frontend): moves TsLink and TsTypography components to their own folder
