+++ components/territorio/link/TS-Link.tsx
@@ -1,9 +1,8 @@
-import TsTypography, { TypographyProps } from './TS-Typography.tsx';
+import TsTypography, { TypographyProps } from "../typography/TS-Typography.tsx";
 
 type LinkProps = TypographyProps & {
-  to: string;
+  to?: string;
   openOnNewTab?: boolean;
-  children: string;
 };
 
 const TsLink = ({
@@ -13,7 +12,7 @@ const TsLink = ({
   ...typographyProps
 }: LinkProps) => {
   return (
-    <a href={to} class='w-fit p-0' target={openOnNewTab ? '_blank' : '_self'}>
+    <a href={to} class="w-fit p-0" target={openOnNewTab ? "_blank" : "_self"}>
       <TsTypography {...typographyProps}>{children}</TsTypography>
     </a>
   );
